Cleaned up code a little bit.

+++ Source/Includes/Types.h
@@ -10,8 +10,8 @@
 	typedef	unsigned	char		uint8;
 	typedef signed		short		sint16;
 	typedef unsigned	short		uint16;
-	typedef signed		int		sint32;
-	typedef unsigned	int		uint32;
+	typedef signed		long		sint32;
+	typedef unsigned	long		uint32;
 	typedef signed		long long	sint64;
 	typedef unsigned	long long	uint64;
